test(dash): Improves coverage
